Split the AwsLibrary and AwsS3Library mixins out into their own files.
